+++ VIva/src/main/java/web/dao/face/CartDao.java
@@ -4,6 +4,11 @@ import java.util.List;
 import java.util.Map;
 
 import web.dto.Cart;
+import web.dto.Credit;
+import web.dto.MySource;
+import web.dto.Source;
+import web.dto.SourceDown;
+import web.dto.Users;
 
 public interface CartDao {
 
@@ -15,6 +20,7 @@ public interface CartDao {
 	 */
 	public List<Map<String, Object>> selectCartByUserno(Cart userNo);
 
+//=========================================================================	
 	/**
 	 * cartNo을 이용하여 cart 항목 삭제하기
 	 * 
@@ -23,7 +29,7 @@ public interface CartDao {
 	 */
 	public boolean deleteCartByCartNo(int cartNo);
 
-
+//=========================================================================	
 	/**
 	 * cart의 중복여부 확인하기. 
 	 * 
@@ -39,6 +45,68 @@ public interface CartDao {
 	 */
 	public void insertCartItem(Cart add);
 
+//=========================================================================	
+	/**
+	 * 회원자격 확인 중, userNo 을 통해 credit TB 조회 하여 총액 구하기
+	 * 
+	 * @param userNo
+	 * @return 크레딧 총계 구하기
+	 */
+	public int selectCreditAcc(Users userNo);
 	
+	/**
+	 * 회원자격 확인 중, 선택한 sourceNo 을 통해 sourcePrice 합계 알아내기
+	 * 
+	 * @param sourceNo
+	 * @return 구매원하는 source 총계 
+	 */
+	public int selectSourceAmount(Source sourceNo);
+	
+	/**
+	 * credit TB에 삽입할 정보(회원번호, 금액)를 이용하여 내역 insert 하기
+	 * 트랜잭션 중, credit TB에 삽입할 정보(회원번호, 금액)를 이용하여 내역 insert 하기
+	 * 
+	 * @param credit - userNo, amount
+	 */
+	public void expenditureCredit(Credit credit);
 
+	/**
+	 * 트랜잭션 중, Cart TB에서 해당 cartNo 항목을 delete 하기
+	 * 
+	 * @param cartNo - cartNo
+	 * 작성자 : 지선
+	 */
+	public void deletePurchasedCartItem(Cart cartNo);
+
+	/**
+	 * 트랜잭션 중, SourceDown TB 에 해당 source 항목 다운로드 내역을 insert 하기
+	 * 
+	 * @param downSource - userNo, sourceNo
+	 * 작성자 : 지선
+	 */
+	public void addSourceToDownList(SourceDown downSource);
+
+	/**
+	 * 트랜잭션 중, 나의음원정보 TB에 기입을 위해
+	 * 구매하려는 음원 정보를 가져온다
+	 * 
+	 * @param sourceNo - 구매하려는 음원소스정보
+	 * @return - 구매음원소스정보 전체
+	 */
+	public Source selectSourceBySourceNo(int sourceNo);
+	
+	/**
+	 * 트랜잭션 중, 사용자가 구매한 음원 정보를 MySource(내가가진음원소스) TB
+	 * 에 추가한다
+	 * 
+	 * @param source - 사용자가 구매한 음원정보
+	 */
+	public void insertMySource(MySource source);
+	
+	/**
+	 * 트랜잭션 중, Credit TB에 구매할 음원의 업로더의 수익금 내역 insert 하기
+	 * 
+	 * @param uploaderInc - userNo, category, amount
+	 */
+	public void uploaderIncomeCredit(Credit uploaderInc);
 }
