+++ VIva/src/main/java/web/dao/mapper/CartMapper.xml
@@ -19,11 +19,13 @@
 		 where C.user_no = #{userNo}
 	</select>
 	
+<!--  ====================================================================== -->	
 	<delete id="deleteCartByCartNo"  parameterType="int">
 		DELETE FROM cart
 		WHERE cart_no = #{cartNo}
 	</delete>
 
+<!--  ====================================================================== -->	
 	<select id="selectDuplicatedCartByUserNo" resultType="int" parameterType="web.dto.Cart">
 		select count(*) from cart
 		where user_no = #{userNo} and source_no = #{sourceNo}
@@ -34,4 +36,68 @@
 		VALUES( cart_seq.nextval, #{sourceNo}, #{userNo})
 	</insert>
 
+<!--  ====================================================================== -->
+	
+	<select id="selectCreditAcc" resultType="int" parameterType="web.dto.Users">
+		SELECT 
+	    (
+	        SELECT nvl(SUM(AMOUNT),0) FROM CREDIT
+	        WHERE DEAL_CATEGORY IN (1, 3)
+	        AND USER_NO=#{userNo}
+	    )
+	    - 
+	    (
+	        SELECT nvl(SUM(AMOUNT),0) FROM CREDIT
+	        WHERE DEAL_CATEGORY IN (2, 4)
+	        AND USER_NO=#{userNo}
+	    ) AMOUNT
+		FROM DUAL
+	</select>
+	
+	<select id="selectSourceAmount"  resultType="int" parameterType="web.dto.Source">
+		SELECT nvl(SUM(SOURCE_PRICE),0) FROM SOURCE
+		WHERE SOURCE_NO IN (${sourceNo})
+	</select>
+
+
+	<!--  트랜잭션 중 -->
+	<!--  트랜잭션 중 1. 회원 크레딧 지출 -->
+	<insert id="expenditureCredit" parameterType="web.dto.Credit">
+		insert into credit (DEAL_NO, user_no, DEAL_CATEGORY, AMOUNT) 
+		values (credit_seq.nextval, #{userNo}, 2, #{amount})
+		<!--  DEAL_DATE 는 현재 DB에서 sysdatd라고 되어있는 관계로...  쿼리문에서는 제외를 해놓은 상태임! -->
+		<!--  DB에서 default CURRENT_TIMESTAMP 라고 지정해놔서 이제 신경 안써도 됨 -->
+	</insert>
+	
+
+	<!--  트랜잭션 중 2. 장바구니 항목에서 삭제 -->
+	<delete id="deletePurchasedCartItem"  parameterType="web.dto.Cart">
+		DELETE FROM cart
+		WHERE cart_no = #{cartNo}
+	</delete>
+	
+	<!--  트랜잭션 중 3. 해당 음원의 다운로드 내역 추가하기 -->
+	<insert id="addSourceToDownList" parameterType="web.dto.SourceDown">
+		INSERT INTO SOURCEDOWN (SOURCEDOWN_NO, USER_NO, SOURCE_NO )
+		VALUES( sourcedown_seq.nextval, #{userNo}, #{sourceNo} )
+	</insert>
+	
+	<!--  트랜잭션 중 4-1. 구매할 음원의 정보 가져오기. (채원 코드 활용 ) -->	
+	<select id="selectSourceBySourceNo" parameterType="int" resultType="web.dto.Source">
+		SELECT * FROM SOURCE
+		WHERE source_no = #{sourceNo}
+	</select>
+	
+	<!--  트랜잭션 중 4-2. 구매할 음원의 정보 MySource TB에 넣기 (채원 코드 활용 ) -->	
+	<insert id="insertMySource" parameterType="web.dto.MySource">
+		INSERT INTO MYSOURCE(LIST_NO, SOURCE_NAME, BPM, KEY, TAG_NO, USER_NO, PACK_NO, SOURCE_NO)
+		VALUES( MYSOURCE_SEQ.NEXTVAL, #{sourceName}, #{bpm}, #{key}, #{tagNo}, #{userNo}, #{packNo}, #{sourceNo})
+	</insert>
+
+	<!--  트랜잭션 중 5. 구매한 음원의 업로더에게 수입내역 insert 하기 -->
+	<insert id="uploaderIncomeCredit" parameterType="web.dto.Credit">
+		insert into credit (DEAL_NO, user_no, DEAL_CATEGORY, AMOUNT) 
+		values (credit_seq.nextval, #{userNo}, #{dealCategory}, #{amount})
+	</insert>
+	
 </mapper>
