+++ VIva/src/main/java/web/service/face/CartService.java
@@ -4,6 +4,8 @@ import java.util.List;
 import java.util.Map;
 
 import web.dto.Cart;
+import web.dto.Source;
+import web.dto.Users;
 
 public interface CartService {
 
@@ -33,6 +35,26 @@ public interface CartService {
 	public boolean addCartItem(Cart add);
 
 
+		
+	/**
+	 * userNo을 통해 크레딧 TB 의 총 계 확인하여 구매 가능여부 확인하기
+	 * 
+	 * @param userNo, sourceNo
+	 * @return 구매 가능 여부
+	 * 작성자 : 지선
+	 */
+	public boolean chkCreditAcc(Users userNo, Source sourceNo);
 	
+	/**
+	 * 크레딧 지출, 장바구니항목 삭제, 다운로드 정보 생성, 내 음원내역 추가, 업로더의 크레딧 수입 생성
+	 * 
+	 * @param userNo - 본인 회원번호
+	 * @param sourceNo - 선택한 음원 항목
+	 * @param cartNo - 해당 장바구니 항목
+	 * @return 구매 완료 여부
+	 * 작성자 : 지선
+	 */
+	public boolean purchaseCartItem(Users userNo, Source sourceNo, Cart cartNo);
+
 
 }
