--- VIva/src/main/java/web/service/face/OrderCartService.java
@@ -1,5 +1,0 @@
-package web.service.face;
-
-public interface OrderCartService {
-
-}
