+++ VIva/src/main/webapp/WEB-INF/views/credit/fail.jsp
@@ -1,0 +1,12 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>Insert title here</title>
+</head>
+<body>
+<h1>실패ㅠㅠ!!</h1>
+</body>
+</html>
